password: extract setPassword/checkPassword to a testable model module
