+++ packages/xxscreeps/mods/backend/password/index.ts
@@ -1,5 +1,5 @@
 import type { Manifest } from 'xxscreeps/config/mods.js';
 
 export const manifest: Manifest = {
-	provides: 'backend',
+	provides: [ 'backend', 'test' ],
 };
